Move everything to ast parsing
We now parse the functions using the ast module. This allows us to not
have to worry about indent, quotes, etc.

We use this to move the code into its own function and add the return
statement for rendering the template.

For extracting the template, we parse the python file again and keep the
final constant as our template. This allows people to define the
template as a triple-quote string at the bottom.