Removes global_objects and adds template environment
The global_objects was a bad solution and prone to breaking. The better
way is to just import what you need from your own app.

We also add in a reference to the template environment on the app
instance, which lets it be modified if needed. (e.g. adding functions or
variables to the globals.)